fix validateAccount duplication
