+++ web/packages/api/src/forInterParachain.ts
@@ -23,6 +23,7 @@ import {
 } from "@polkadot/types/interfaces"
 import { ParachainBase } from "./parachains/parachainBase"
 import { u8aToHex } from "@polkadot/util"
+import { validateAccount } from "./assets_v2"
 
 export type Transfer = {
     input: {
@@ -471,30 +472,6 @@ export async function signAndSend(
     result.messageId = transfer.computed.messageId ?? result.messageId
 
     return result
-}
-
-async function validateAccount(
-    parachainImpl: ParachainBase,
-    beneficiaryAddress: string,
-    ethChainId: number,
-    tokenAddress: string,
-    assetMetadata?: Asset,
-    maxConsumers?: bigint
-) {
-    // Check if the account is created
-    const [beneficiaryAccount, beneficiaryTokenBalance] = await Promise.all([
-        parachainImpl.getNativeAccount(beneficiaryAddress),
-        parachainImpl.getTokenBalance(beneficiaryAddress, ethChainId, tokenAddress, assetMetadata),
-    ])
-    return {
-        accountExists: !(
-            beneficiaryAccount.consumers === 0n &&
-            beneficiaryAccount.providers === 0n &&
-            beneficiaryAccount.sufficients === 0n
-        ),
-        accountMaxConsumers:
-            beneficiaryAccount.consumers >= (maxConsumers ?? 63n) && beneficiaryTokenBalance === 0n,
-    }
 }
 
 function createTx(
