+++ web/packages/api/src/transfers/toPolkadot/pnaToAH.ts
@@ -14,13 +14,17 @@ import {
     encodeNativeAsset,
     hexToBytes,
     Transfer,
-    validateAccount,
     ValidationKind,
     ValidationResult,
 } from "../../toPolkadotSnowbridgeV2"
 import { accountId32Location, DOT_LOCATION, erc20Location } from "../../xcmBuilder"
 import { paraImplementation } from "../../parachains"
-import { erc20Balance, ETHER_TOKEN_ADDRESS, swapAsset1ForAsset2 } from "../../assets_v2"
+import {
+    erc20Balance,
+    ETHER_TOKEN_ADDRESS,
+    swapAsset1ForAsset2,
+    validateAccount,
+} from "../../assets_v2"
 import { beneficiaryMultiAddress, padFeeByPercentage } from "../../utils"
 import { FeeInfo, resolveInputs, ValidationLog, ValidationReason } from "../../toPolkadot_v2"
 import { Contract } from "ethers"
@@ -337,7 +341,7 @@ export class PNAToAH implements TransferInterface {
         const assetHubImpl = await paraImplementation(assetHub)
 
         if (!ahAssetMetadata.isSufficient && !dryRunAhSuccess) {
-            const { accountMaxConumers, accountExists } = await validateAccount(
+            const { accountMaxConsumers, accountExists } = await validateAccount(
                 assetHubImpl,
                 beneficiaryAddressHex,
                 registry.ethChainId,
@@ -345,7 +349,7 @@ export class PNAToAH implements TransferInterface {
                 ahAssetMetadata
             )
 
-            if (accountMaxConumers) {
+            if (accountMaxConsumers) {
                 logs.push({
                     kind: ValidationKind.Error,
                     reason: ValidationReason.MaxConsumersReached,
