+++ web/packages/api/src/transfers/toPolkadot/pnaToParachain.ts
@@ -15,13 +15,17 @@ import {
     encodeNativeAsset,
     hexToBytes,
     Transfer,
-    validateAccount,
     ValidationKind,
     ValidationResult,
 } from "../../toPolkadotSnowbridgeV2"
 import { accountId32Location, DOT_LOCATION, erc20Location } from "../../xcmBuilder"
 import { paraImplementation } from "../../parachains"
-import { erc20Balance, ETHER_TOKEN_ADDRESS, swapAsset1ForAsset2 } from "../../assets_v2"
+import {
+    erc20Balance,
+    ETHER_TOKEN_ADDRESS,
+    swapAsset1ForAsset2,
+    validateAccount,
+} from "../../assets_v2"
 import { beneficiaryMultiAddress, padFeeByPercentage, paraIdToSovereignAccount } from "../../utils"
 import { FeeInfo, resolveInputs, ValidationLog, ValidationReason } from "../../toPolkadot_v2"
 import {
@@ -413,7 +417,7 @@ export class PNAToParachain implements TransferInterface {
         // Check if sovereign account balance for token is at 0 and that consumers is maxxed out.
         if (!ahAssetMetadata.isSufficient && !dryRunAhSuccess) {
             const sovereignAccountId = paraIdToSovereignAccount("sibl", destinationParaId)
-            const { accountMaxConumers, accountExists } = await validateAccount(
+            const { accountMaxConsumers, accountExists } = await validateAccount(
                 assetHubImpl,
                 sovereignAccountId,
                 registry.ethChainId,
@@ -428,7 +432,7 @@ export class PNAToParachain implements TransferInterface {
                     message: "Sovereign account does not exist on Asset Hub.",
                 })
             }
-            if (accountMaxConumers) {
+            if (accountMaxConsumers) {
                 logs.push({
                     kind: ValidationKind.Error,
                     reason: ValidationReason.MaxConsumersReached,
@@ -489,14 +493,14 @@ export class PNAToParachain implements TransferInterface {
             ) {
                 const destParachainImpl = await paraImplementation(destParachainApi)
                 // Check if the account is created
-                const { accountMaxConumers, accountExists } = await validateAccount(
+                const { accountMaxConsumers, accountExists } = await validateAccount(
                     destParachainImpl,
                     beneficiaryAddressHex,
                     registry.ethChainId,
                     tokenAddress,
                     destAssetMetadata
                 )
-                if (accountMaxConumers) {
+                if (accountMaxConsumers) {
                     logs.push({
                         kind: ValidationKind.Error,
                         reason: ValidationReason.MaxConsumersReached,
